tests/inputdevicecapabilities-ime.html

<!DOCTYPE html>
<title>InputDeviceCapabilities test for keyboard-triggered events</title>
<meta name="viewport" content="initial-scale=1">
<meta charset="UTF-8">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevicecapabilities-tests.js"></script>
<style>
#target {
  margin-bottom: 20px;
}
</style>

<h1>InputDeviceCapabilities - test for IME-triggered events</h1>
<p>Enter text into the input field using an input method editor (eg. virtual keyboard).
</p>
<input id="target">

<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceCapabilitiesValidator(sourceCapabilities) {
  assert_true(sourceCapabilities != null, "sourceCapabilities is set");
  assert_false(sourceCapabilities.firesTouchEvents, "sourceCapabilities.firesTouchEvents");
}

function runTests() {
  var events = ["keydown", "keyup", "keypress", "compositionstart", "compositionupdate", "compositionend"];

  // Some browsers don't yet implement 'input' as a UIEvent, and many browsers don't yet support
  // 'beforeinput'
  if ("InputEvent" in window) {
    if ("oninput" in document.body)
      events.push("input");
    if ("onbeforeinput" in document.body)
      events.push("beforeinput");
  }

  runInputDeviceEventTests(events, inputDeviceValidator);
}
</script>
